#15 replaced useStates with useRefs for input forms and deleted PersonalSetting.tsx
